FavIcon & Homepage App Metadata data
Used: https://realfavicongenerator.net

Figure this will be on mobile devices so it's good to have a nice icon and config for it we can customize to look good offline and as a local app on tablets. Will need to tweak as we go but this is a start. Minor thing but needed later.